DownloadDash-API/DownnloadDash-API/requirements-full.txt
# Full dependency set (optional)
#
# Note: some packages in this list may require extra system dependencies or have
# limited wheel availability on very new Python versions. Use this when you
# need the optional features.

# Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
# psycopg2-binary does not currently provide wheels for all Python 3.13 builds.
# Prefer psycopg (v3) for Postgres:
psycopg[binary]>=3.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx-oauth==0.4.0

# Media Downloading
yt-dlp==2023.10.13
youtube-dl==2021.12.17
pytube==15.0.0
instaloader==4.10.2
instagrapi==2.1.3
social-media-downloader==0.2.1

# TikTok Content Scraper (optional, unofficial)
git+https://github.com/Q-Bukold/TikTok-Content-Scraper.git

# Task Queue
celery==5.3.4
redis==5.0.1

# File Handling
aiofiles==23.2.1
pillow==10.1.0
moviepy==1.0.3

# Networking / Parsing
httpx==0.25.1
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
python-magic==0.4.27

# WebSocket
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
